+++ lib/main_db/new/providers/index.dart
@@ -1,4 +1,4 @@
 library;
 
 export 'db_history_provider.dart';
-export 'main_store_manager_provider.dart';
+export 'main_store_service_provider.dart';